Implemented Types Header & Added _micros asserts
