+++ src/ArduinoMCP2515.h
@@ -18,12 +18,12 @@
 #include "MCP2515/MCP2515_Io.h"
 #include "MCP2515/MCP2515_Config.h"
 #include "MCP2515/MCP2515_Control.h"
+#include "MCP2515/MCP2515_Types.h"
 
 #undef min
 #undef max
 #include <vector>
 #include <string>
-#include <functional>
 
 #if defined __has_include
 #  if __has_include (<libcanard/canard.h>)
@@ -81,17 +81,6 @@ enum class CanBitRate : size_t
   BR_500kBPS_12MHZ,
   BR_1000kBPS_12MHZ
 };
-
-typedef std::function<unsigned long()> MicroSecondFunc;
-#if LIBCANARD
-typedef std::function<void(CanardFrame const & frame)> OnReceiveBufferFullFunc;
-#else
-typedef std::function<void(uint32_t const, uint32_t const, uint8_t const *, uint8_t const)> OnReceiveBufferFullFunc;
-#endif
-class ArduinoMCP2515;
-typedef std::function<void(ArduinoMCP2515 *)> OnTransmitBufferEmptyFunc;
-typedef std::function<void(MCP2515::EFLG const)> OnCanErrorFunc;
-typedef std::function<void(MCP2515::EFLG const)> OnCanWarningFunc;
 
 /**************************************************************************************
  * CLASS DECLARATION
