+++ src/MCP2515/MCP2515_Config.cpp
@@ -9,6 +9,8 @@
  * INCLUDE
  **************************************************************************************/
 
+#include <assert.h>
+
 #include "MCP2515_Config.h"
 
 /**************************************************************************************
@@ -26,7 +28,7 @@ MCP2515_Config::MCP2515_Config(MCP2515_Io & io, MicroSecondFunc micros)
 : _io{io}
 , _micros{micros}
 {
-
+  assert(_micros != nullptr);
 }
 
 /**************************************************************************************
